+++ client/src/components/ErrorCards/ErrorCard.jsx
@@ -16,7 +16,6 @@ const useStyles = makeStyles((theme) => ({
     flexDirection: 'column',
     alignItems: 'center',
     justifyContent: 'center',
-    // backgroundColor: red[100],
     backgroundColor: theme.palette.background.error,
     color: theme.palette.getContrastText(theme.palette.background.error),
     width: '100%',
@@ -28,7 +27,6 @@ const useStyles = makeStyles((theme) => ({
       duration: theme.transitions.duration.short,
     }),
     '&:hover': {
-      // backgroundColor: red[200],
       backgroundColor: lighten(theme.palette.background.error, 0.3),
     },
     '&:focus, &:active': {
@@ -51,11 +49,11 @@ const useStyles = makeStyles((theme) => ({
 /**
  * A card representing that there was an error getting the stats
  */
-const StatsErrorCard = ({ fetchStatsCompare, className }) => {
+const ErrorCard = ({ retryFunc, className }) => {
   const classes = useStyles();
 
   const handleClick = () => {
-    fetchStatsCompare();
+    retryFunc();
   };
 
   return (
@@ -74,20 +72,4 @@ const StatsErrorCard = ({ fetchStatsCompare, className }) => {
   );
 };
 
-StatsErrorCard.defaultProps = {
-  className: null,
-};
-
-StatsErrorCard.propTypes = {
-  /** A function used to fetch the stats */
-  fetchStatsCompare: PropTypes.func.isRequired,
-  /** CSS classname to give the component */
-  className: PropTypes.string,
-};
-
-
-const mapDispatchToProps = (dispatch) => bindActionCreators({
-  fetchStatsCompare,
-}, dispatch);
-
-export default connect(null, mapDispatchToProps)(StatsErrorCard);
+export default ErrorCard;
