+++ client/src/components/ErrorCards/stories/StatsErrorCard.story.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
-import StatsErrorCard from 'components/StatsErrorCard';
+import { StatsErrorCard } from 'components/ErrorCards';
 import { withProvider } from 'utils/exampleStore';
 
 storiesOf('Components/StatsErrorCard', module)
