--- client/src/components/StatsErrorCard/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './StatsErrorCard';
